refactored terminal cli and made view leaderboards sort
